Don't store shared_ptr of parent: fix memleaks
Previously tree nodes were never deleted because children hold their parents and vice versa.
To properly free memory when Tlv elements go out of scope, nodes cannot hold a shared_ptr of
their parent.

Instead store a raw pointer to not loose the convenience of detach() and has_parent()
methods. But we can't allow external access to the parent via child because the child does not
havea shared_ptr of the parent.

This chould be overcome by storing a weak_ptr or derving Data from std::enable_shared_from_this
but it's probably not worth the overhead.